Add AuthzExtension to get information from headers
Add AuthzExtension to get principal and
group record information from headers
passed from external IDP.

When integration with external IDP is
enabled in engine we can configure misc
AuthzExtension to get the principal and
groups records from the AuthzExtension.

Authz search for specific groups and users
returns the specifc user or group. If the
search contains any wild card an empty result
is returned.

Query to fetch the principal record uses the
principal name and groups name configured in
the extensions property file to get the principal
name and the groups from the OpenID connect claim
set.

The README.http file contains the details
of how to configure the Authz extension.

Change-Id: I20161617c97e112b3aec891296cea4c70c6a7d05
Bug-Url: https://bugzilla.redhat.com/1570040
Signed-off-by: Ravi Nori <rnori@redhat.com>